m
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/StatisticalTests/Trunk@44528 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/statisticaltests/test/CorrelationTest.java
@@ -1,15 +1,21 @@
 package org.processmining.statisticaltests.test;
 
+import java.awt.image.BufferedImage;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
+import java.io.IOException;
 import java.math.BigDecimal;
+
+import javax.imageio.ImageIO;
 
 import org.deckfour.xes.classification.XEventNameClassifier;
 import org.deckfour.xes.model.XLog;
 import org.processmining.correlation.CorrelationParametersAbstract;
+import org.processmining.correlation.CorrelationPlot;
+import org.processmining.correlation.CorrelationPlotLegend;
 import org.processmining.correlation.CorrelationProcessNumerical;
 import org.processmining.framework.plugin.PluginContext;
 import org.processmining.framework.plugin.ProMCanceller;
@@ -19,43 +25,149 @@ import org.processmining.plugins.inductiveminer2.attributes.AttributeImpl;
 import org.processmining.plugins.inductiveminer2.attributes.AttributeImpl.Type;
 import org.processmining.xeslite.plugin.OpenLogFileLiteImplPlugin;
 
+import gnu.trove.list.TDoubleList;
+import gnu.trove.list.array.TDoubleArrayList;
+
 public class CorrelationTest {
 	static File folder = new File("/home/sander/Documents/svn/41 - stochastic statistics/experiments/logs");
 
-	static int maxSampleSize = 150000;
-	static int numberOfSamples = 10000;
-
 	public static void main(String... args) throws Exception {
-		testLogSingle();
+//		bpic20DDAmountSingle();
+		bpic20DDAmountPlot();
+
+		//		testLogSingle(1);
+		//		testLogPlot();
+		//		bpic11AgePlot();
+		//		bpic12aAmountRequestedPlot();
+		//		roadFinesAmountRequestedPlot();
+	}
+
+	public static void legend() throws IOException {
+		File outputFile = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/legend.png");
+		BufferedImage image = new CorrelationPlotLegend().create("sample density", "1", "highest");
+		ImageIO.write(image, "png", outputFile);
 	}
 
 	public static void testLog() throws Exception {
 		File inputLog = new File(folder, "correlation-test-log.xes.gz");
 		Attribute attribute = new AttributeImpl("value", Type.numeric);
 		File outputCsv = new File(
-				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/testLog-value-samsen-1000.csv");
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/testLog-value-samsen.csv");
 
-		multipleCorrelation(inputLog, outputCsv, 1000, attribute);
+		multipleCorrelation(inputLog, outputCsv, 100, 100, attribute);
 	}
 
-	public static void testLogSingle() throws Exception {
-		int testLog = 2;
+	public static void testLogSingle(int testLog) throws Exception {
 		File inputLog = new File(folder, "testLog" + testLog + ".xes.gz");
 		Attribute attribute = new AttributeImpl("value", Type.numeric);
 		File outputCsv = new File(
 				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/testLog" + testLog
 						+ ".csv");
 
-		correlation(inputLog, outputCsv, attribute, 500);
+		correlation(inputLog, outputCsv, attribute, 999500, true);
+	}
+
+	public static void testLogPlot() throws Exception {
+		for (int testLog = 1; testLog <= 2; testLog++) {
+			File outputCsv = new File(
+					"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/testLog"
+							+ testLog + ".csv");
+			File outputImageFile = new File(
+					"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/testLog"
+							+ testLog + ".png");
+
+			createCorrelationPlot(outputCsv, outputImageFile);
+		}
+	}
+
+	public static void bpic20DDAmountSingle() throws Exception {
+		File inputLog = new File(folder, "bpic20-DomesticDeclarations.xes.gz");
+		Attribute attribute = new AttributeImpl("Amount", Type.numeric);
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic20-DomesticDeclarations-Amount.csv");
+
+		correlation(inputLog, outputCsv, attribute, 1000000, true);
+	}
+	
+	public static void bpic20DDAmountPlot() throws Exception {
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic20-DomesticDeclarations-Amount.csv");
+		File outputImageFile = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic20-DomesticDeclarations-Amount.png");
+
+		createCorrelationPlot(outputCsv, outputImageFile);
+	}
+
+	public static void bpic15EndDate(int municipality) throws Exception {
+		File inputLog = new File(folder, "BPIC15_" + municipality + ".xes");
+		Attribute attribute = new AttributeImpl("endDate", Type.time);
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/BPIC15_"
+						+ municipality + "-endDate-samsen.csv");
+
+		multipleCorrelation(inputLog, outputCsv, 10000, 1000000, attribute);
+	}
+
+	public static void bpic15EndDateSingle(int municipality) throws Exception {
+		File inputLog = new File(folder, "BPIC15_" + municipality + ".xes");
+		Attribute attribute = new AttributeImpl("endDate", Type.time);
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/BPIC15_"
+						+ municipality + "-endDate.csv");
+
+		correlation(inputLog, outputCsv, attribute, 1000000, true);
+	}
+
+	public static void bpic15EndDatePlot(int municipality) throws Exception {
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/BPIC15_"
+						+ municipality + "-endDate.csv");
+		File outputImageFile = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/BPIC15_"
+						+ municipality + "-endDate.png");
+
+		createCorrelationPlot(outputCsv, outputImageFile);
+	}
+
+	public static void bpic15SumLeges(int municipality) throws Exception {
+		File inputLog = new File(folder, "BPIC15_" + municipality + ".xes");
+		Attribute attribute = new AttributeImpl("SUMleges", Type.numeric);
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/BPIC15_"
+						+ municipality + "-SUMleges-samsen.csv");
+
+		multipleCorrelation(inputLog, outputCsv, 10000, 1000000, attribute);
+	}
+
+	public static void bpic15SumLegesSingle(int municipality) throws Exception {
+		File inputLog = new File(folder, "BPIC15_" + municipality + ".xes");
+		Attribute attribute = new AttributeImpl("SUMleges", Type.numeric);
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/BPIC15_"
+						+ municipality + "-SUMLeges.csv");
+
+		correlation(inputLog, outputCsv, attribute, 1000000, true);
+	}
+
+	public static void bpic15SumLegesPlot(int municipality) throws Exception {
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/BPIC15_"
+						+ municipality + "-SUMLeges.csv");
+		File outputImageFile = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/BPIC15_"
+						+ municipality + "-SUMLeges.png");
+
+		createCorrelationPlot(outputCsv, outputImageFile);
 	}
 
 	public static void bpic11Age() throws Exception {
 		File inputLog = new File(folder, "bpic11.xes.gz");
 		Attribute attribute = new AttributeImpl("Age:1", Type.numeric);
 		File outputCsv = new File(
-				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/bpic11-Age1-samsen-100.csv");
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/bpic11-Age1-samsen.csv");
 
-		multipleCorrelation(inputLog, outputCsv, 100, attribute);
+		multipleCorrelation(inputLog, outputCsv, 10000, 1000000, attribute);
 	}
 
 	public static void bpic11AgeSingle() throws Exception {
@@ -64,16 +176,25 @@ public class CorrelationTest {
 		File outputCsv = new File(
 				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic11-Age1.csv");
 
-		correlation(inputLog, outputCsv, attribute, 500);
+		correlation(inputLog, outputCsv, attribute, 998000, true);
+	}
+
+	public static void bpic11AgePlot() throws Exception {
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic11-Age1.csv");
+		File image = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic11-Age1.png");
+
+		createCorrelationPlot(outputCsv, image);
 	}
 
 	public static void bpic12aAmountRequested() throws Exception {
 		File inputLog = new File(folder, "bpic12-a.xes");
 		Attribute attribute = new AttributeImpl("AMOUNT_REQ", Type.numeric);
 		File outputCsv = new File(
-				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/bpic12a-AMOUNT_REQ-samsen-1000.csv");
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/bpic12a-AMOUNT_REQ-samsen.csv");
 
-		multipleCorrelation(inputLog, outputCsv, 1000, attribute);
+		multipleCorrelation(inputLog, outputCsv, 10000, 1000000, attribute);
 	}
 
 	public static void bpic12aAmountRequestedSingle() throws Exception {
@@ -82,16 +203,61 @@ public class CorrelationTest {
 		File outputCsv = new File(
 				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic12a-AMOUNT_REQ.csv");
 
-		correlation(inputLog, outputCsv, attribute, 500);
+		correlation(inputLog, outputCsv, attribute, 700000, true);
+	}
+
+	public static void bpic12aAmountRequestedPlot() throws Exception {
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic12a-AMOUNT_REQ.csv");
+		File outputImageFile = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic12a-AMOUNT_REQ.png");
+
+		createCorrelationPlot(outputCsv, outputImageFile);
+	}
+
+	public static void bpic12AmountRequestedSingle() throws Exception {
+		File inputLog = new File(folder, "bpic12.xes.gz");
+		Attribute attribute = new AttributeImpl("AMOUNT_REQ", Type.numeric);
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic12-AMOUNT_REQ.csv");
+
+		correlation(inputLog, outputCsv, attribute, 1000000, true);
+	}
+
+	public static void bpic12AmountRequestedPlot() throws Exception {
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic12-AMOUNT_REQ.csv");
+		File outputImageFile = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic12-AMOUNT_REQ.png");
+
+		createCorrelationPlot(outputCsv, outputImageFile);
+	}
+
+	public static void bpic17AmountRequestedSingle() throws Exception {
+		File inputLog = new File(folder, "BPI Challenge 2017.xes.gz");
+		Attribute attribute = new AttributeImpl("RequestedAmount", Type.numeric);
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic17-RequestedAmount.csv");
+
+		correlation(inputLog, outputCsv, attribute, 1000000, true);
+	}
+
+	public static void bpic17AmountRequestedPlot() throws Exception {
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic17-RequestedAmount.csv");
+		File outputImageFile = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/bpic17-RequestedAmount.png");
+
+		createCorrelationPlot(outputCsv, outputImageFile);
 	}
 
 	public static void roadFinesAmountRequested() throws Exception {
 		File inputLog = new File(folder, "Road fines with trace attributes.xes.gz");
 		Attribute attribute = new AttributeImpl("amount", Type.numeric);
 		File outputCsv = new File(
-				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/roadFines-amount-samsen-100.csv");
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/05 - correlation sampleSize sensitivity/roadFines-amount-samsen.csv");
 
-		multipleCorrelation(inputLog, outputCsv, 100, attribute);
+		multipleCorrelation(inputLog, outputCsv, 10000, 1000000, attribute);
 	}
 
 	public static void roadFinesAmountRequestedSingle() throws Exception {
@@ -100,15 +266,55 @@ public class CorrelationTest {
 		File outputCsv = new File(
 				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/roadFines-amount.csv");
 
-		correlation(inputLog, outputCsv, attribute, 500);
+		correlation(inputLog, outputCsv, attribute, 980000, true);
 	}
 
-	private static void correlation(File inputLog, File outputCsv, Attribute attribute, int numberOfSamples)
-			throws Exception {
+	public static void roadFinesAmountRequestedPlot() throws Exception {
+		File outputCsv = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/roadFines-amount.csv");
+		File image = new File(
+				"/home/sander/Documents/svn/41 - stochastic statistics/experiments/04 - correlation/roadFines-amount.png");
+
+		createCorrelationPlot(outputCsv, image);
+	}
+
+	public static void createCorrelationPlot(File fileCsv, File outputImageFile) throws IOException {
+		CorrelationPlot plot = new CorrelationPlot();
+
+		//read the correlation sample file
+		TDoubleList deltaValues = new TDoubleArrayList();
+		TDoubleList deltaTraces = new TDoubleArrayList();
+		BufferedReader reader = new BufferedReader(new FileReader(fileCsv));
+		reader.readLine();
+		String line = reader.readLine();
+		while (line != null) {
+			String[] arr = line.split(",");
+			deltaValues.add(Double.valueOf(arr[0]));
+			deltaTraces.add(Double.valueOf(arr[1]));
+
+			line = reader.readLine();
+		}
+		reader.close();
+
+		System.out.println("number of samples " + deltaValues.size());
+
+		outputImageFile.mkdirs();
+
+		BufferedImage image = plot.create("Δ value", deltaValues.toArray(), "Δ trace", deltaTraces.toArray());
+		ImageIO.write(image, "png", outputImageFile);
+	}
+
+	private static void correlation(File inputLog, File outputCsv, Attribute attribute, int numberOfSamples,
+			boolean append) throws Exception {
 		outputCsv.getParentFile().mkdirs();
-		outputCsv.createNewFile();
-		BufferedWriter output = new BufferedWriter(new FileWriter(outputCsv, false));
-		output.write("valueDelta,processDelta\n");
+		BufferedWriter output;
+		if (!append || !outputCsv.exists()) {
+			outputCsv.createNewFile();
+			output = new BufferedWriter(new FileWriter(outputCsv, false));
+			output.write("valueDelta,processDelta\n");
+		} else {
+			output = new BufferedWriter(new FileWriter(outputCsv, append));
+		}
 
 		PluginContext context = new FakeContext();
 		XLog log = (XLog) new OpenLogFileLiteImplPlugin().importFile(context, inputLog);
@@ -130,25 +336,33 @@ public class CorrelationTest {
 		}
 		output.flush();
 		output.close();
+
+		double[] x = result[0];
+		double[] y = result[1];
+		BigDecimal meanY = Correlation.mean(y);
+		double standardDeviationYd = Correlation.standardDeviation(y, meanY);
+		double correlation = Correlation.correlation(x, y, meanY, standardDeviationYd).doubleValue();
+
+		System.out.println("correlation " + correlation);
 	}
 
-	private static void multipleCorrelation(File inputLog, File outputCsv, int step, Attribute attribute)
-			throws Exception {
+	private static void multipleCorrelation(File inputLog, File outputCsv, int step, int maxNumberOfSamples,
+			Attribute attribute) throws Exception {
 		outputCsv.getParentFile().mkdirs();
-		int startSampleSize;
+		int startNumberOfSamples;
 		BufferedWriter output;
 		if (!outputCsv.exists()) {
 			outputCsv.createNewFile();
-			startSampleSize = step;
+			startNumberOfSamples = step;
 			output = new BufferedWriter(new FileWriter(outputCsv, false));
-			output.write("sampleSize,correlation,time\n");
+			output.write("numberOfSamples,correlation,time\n");
 		} else {
 
 			//count the number of lines in the file
 			BufferedReader reader = new BufferedReader(new FileReader(outputCsv));
-			startSampleSize = 0;
+			startNumberOfSamples = 0;
 			while (reader.readLine() != null) {
-				startSampleSize += step;
+				startNumberOfSamples += step;
 			}
 			reader.close();
 
@@ -164,12 +378,12 @@ public class CorrelationTest {
 			}
 		};
 
-		CorrelationParametersAbstract parameters = new CorrelationParametersAbstract(numberOfSamples,
-				new XEventNameClassifier(), attribute, System.currentTimeMillis(), true) {
+		CorrelationParametersAbstract parameters = new CorrelationParametersAbstract(0, new XEventNameClassifier(),
+				attribute, System.currentTimeMillis(), true) {
 		};
 
-		for (int sampleSize = startSampleSize; sampleSize <= maxSampleSize; sampleSize += step) {
-			System.out.println("sample size " + sampleSize);
+		for (int numberOfSamples = startNumberOfSamples; numberOfSamples <= maxNumberOfSamples; numberOfSamples += step) {
+			parameters.setNumberOfSamples(numberOfSamples);
 
 			long startTime = System.currentTimeMillis();
 			double[][] result = CorrelationProcessNumerical.compute(parameters, log, canceller);
@@ -182,7 +396,7 @@ public class CorrelationTest {
 			double standardDeviationYd = Correlation.standardDeviation(y, meanY);
 			double correlation = Correlation.correlation(x, y, meanY, standardDeviationYd).doubleValue();
 
-			output.write(sampleSize + "," + correlation + "," + time + "\n");
+			output.write(numberOfSamples + "," + correlation + "," + time + "\n");
 			output.flush();
 		}
 		output.close();
